Rename splitting package to partitioning
