+++ java/system-test/system-test-drivers/src/main/java/sleeper/systemtest/drivers/compaction/SplitPartitionsUntilNoMoreSplits.java
@@ -32,7 +32,7 @@ import sleeper.core.statestore.StateStore;
 import sleeper.core.statestore.StateStoreException;
 import sleeper.statestore.StateStoreProvider;
 import sleeper.systemtest.configuration.SystemTestProperties;
-import sleeper.systemtest.drivers.splitting.WaitForPartitionSplitting;
+import sleeper.systemtest.drivers.partitioning.WaitForPartitionSplitting;
 import sleeper.systemtest.drivers.util.InvokeSystemTestLambda;
 
 import java.io.IOException;
