+++ java/system-test/system-test-drivers/src/main/java/sleeper/systemtest/drivers/partitioning/PartitionSplittingDriver.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package sleeper.systemtest.drivers.splitting;
+package sleeper.systemtest.drivers.partitioning;
 
 import software.amazon.awssdk.services.lambda.LambdaClient;
