+++ java/system-test/system-test-drivers/src/main/java/sleeper/systemtest/drivers/partitioning/WaitForPartitionSplitting.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package sleeper.systemtest.drivers.splitting;
+package sleeper.systemtest.drivers.partitioning;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
